Add cast for printf.

+++ clients/odbc/driver/SQLGetConnectAttr.c
@@ -201,7 +201,8 @@ SQLGetConnectAttrW(SQLHDBC ConnectionHandle,
 	ODBCLOG("SQLGetConnectAttrW " PTRFMT " %s " PTRFMT " %d " PTRFMT "\n",
 		PTRFMTCAST ConnectionHandle,
 		translateConnectAttribute(Attribute),
-		PTRFMTCAST ValuePtr, BufferLength, PTRFMTCAST StringLengthPtr);
+		PTRFMTCAST ValuePtr, (int) BufferLength,
+		PTRFMTCAST StringLengthPtr);
 #endif
 
 	if (!isValidDbc(dbc))
